Fix WebAuthn login
